SysDept增加deptTyep字段; 小区基本信息录入页面; 所属小区选择组件;

+++ station-server/sps-system/src/main/java/com/jingyu/community/mapper/CommunityDescMapper.java
@@ -57,7 +57,7 @@ public interface CommunityDescMapper {
      */
     public int deleteSqCommunityDescByIds(Long[] ids);
 
-    void updateCommunityId(@Param("id") int id,@Param("communityId") Long communityId);
+    void updateCommunityId(@Param("id") long id,@Param("communityId") Long communityId);
 
     void deleteByCommunityId(Long id);
 }
